refactor(compiler): Make ElementAttributes an implementation detail
Refactors ElementAttributes to be an implementation detail of the const
collection phase, rather than an object that is added to all ElementOps.
