+++ packages/compiler/src/template/pipeline/ir/index.ts
@@ -6,9 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export * from './src/element';
-export * from './src/expression';
 export * from './src/enums';
+export * from './src/expression';
 export * from './src/operations';
 export * from './src/ops/create';
 export * from './src/ops/host';
